Added return shared element fragment animation
This took ages to work out!? The key was need to postponeEnterTransition in the first Fragment. Think this is because the back fragment creates view again, unlike activity? Postponing caused a lot of problems because need to set  a transitioner on it so that it can startPostponed when ready.
The remapping was a nightmare too. This article and attached code helped loads, https://android-developers.googleblog.com/2018/02/continuous-shared-element-transitions.html. It's not much like Activities at all - the remapping especially. Think because Fragments share an Activity so can talk to each other
