+++ NavigationReactNative/src/android/app/src/main/java/com/navigation/reactnative/ActivityNavigator.java
@@ -6,6 +6,15 @@ import android.content.Intent;
 import android.os.Build;
 import android.os.Bundle;
 import android.util.Pair;
+import android.view.View;
+
+import androidx.core.app.SharedElementCallback;
+
+import com.facebook.react.views.image.ReactImageView;
+
+import java.util.HashMap;
+import java.util.List;
+import java.util.Map;
 
 class ActivityNavigator extends SceneNavigator {
     private static final String ORIENTATION = "Navigation.ORIENTATION";
@@ -67,5 +76,58 @@ class ActivityNavigator extends SceneNavigator {
         activity.finish();
         activity.startActivity(intent);
         activity.overridePendingTransition(enter, exit);
+    }
+
+    private Pair[] getOldSharedElements(int currentCrumb, int crumb, SharedElementContainer sharedElementContainer, final NavigationStackView stack, final Activity activity) {
+        final HashMap<String, View> oldSharedElementsMap = getSharedElementMap(sharedElementContainer.getScene());
+        final Pair[] oldSharedElements = currentCrumb - crumb == 1 ? getSharedElements(oldSharedElementsMap, stack.oldSharedElementNames) : null;
+        if (oldSharedElements != null && oldSharedElements.length != 0) {
+            final SharedElementTransitioner transitioner = new SharedElementTransitioner(sharedElementContainer, getSharedElementSet(stack.oldSharedElementNames));
+            sharedElementContainer.setEnterCallback(new SharedElementCallback() {
+                @Override
+                public void onMapSharedElements(List<String> names, Map<String, View> elements) {
+                    names.clear();
+                    elements.clear();
+                    for(int i = 0; i < stack.oldSharedElementNames.size(); i++) {
+                        String name = stack.oldSharedElementNames.getString(i);
+                        names.add(name);
+                        if (oldSharedElementsMap.containsKey(name)) {
+                            View oldSharedElement = oldSharedElementsMap.get(name);
+                            elements.put(name, oldSharedElement);
+                            SharedElementView oldSharedElementView = (SharedElementView) oldSharedElement.getParent();
+                            transitioner.load(name, oldSharedElementView.exitTransition, activity);
+                        }
+                    }
+                }
+            });
+            return oldSharedElements;
+        }
+        return null;
+    }
+
+    private Pair[] getSharedElements(int currentCrumb, int crumb, SharedElementContainer sharedElementContainer, NavigationStackView stack) {
+        final HashMap<String, View> sharedElementsMap = getSharedElementMap(sharedElementContainer.getScene());
+        final Pair[] sharedElements = crumb - currentCrumb == 1 ? getSharedElements(sharedElementsMap, stack.sharedElementNames) : null;
+        if (sharedElements != null && sharedElements.length != 0) {
+            sharedElementContainer.setExitCallback(new SharedElementCallback() {
+                @Override
+                public void onSharedElementEnd(List<String> names, List<View> elements, List<View> snapshots) {
+                    for (View view : elements) {
+                        if (view instanceof ReactImageView)
+                            ((ReactImageView) view).getDrawable().setVisible(true, true);
+                    }
+                }
+                @Override
+                public void onMapSharedElements(List<String> names, Map<String, View> elements) {
+                    elements.clear();
+                    for(String name : names) {
+                        if (sharedElementsMap.containsKey(name))
+                            elements.put(name, sharedElementsMap.get(name));
+                    }
+                }
+            });
+            return sharedElements;
+        }
+        return null;
     }
 }
