+++ NavigationReactNative/src/android/app/src/main/java/com/navigation/reactnative/FragmentNavigator.java
@@ -4,15 +4,25 @@ import android.app.Activity;
 import android.util.Pair;
 import android.view.View;
 
+import androidx.core.app.SharedElementCallback;
 import androidx.fragment.app.FragmentActivity;
 import androidx.fragment.app.FragmentManager;
 import androidx.fragment.app.FragmentTransaction;
+
+import java.util.HashMap;
+import java.util.List;
+import java.util.Map;
 
 class FragmentNavigator extends SceneNavigator {
 
     @Override
     void navigateBack(int currentCrumb, int crumb, Activity activity, NavigationStackView stack) {
         FragmentManager fragmentManager = ((FragmentActivity) activity).getSupportFragmentManager();
+        SceneFragment fragment = (SceneFragment) fragmentManager.getFragments().get(fragmentManager.getFragments().size() - 1);
+        Pair[] sharedElements = getOldSharedElements(currentCrumb, crumb, fragment, stack);
+        SceneFragment prevFragment = (SceneFragment) fragmentManager.findFragmentByTag(stack.keys.getString(crumb));
+        if (sharedElements != null && prevFragment != null && prevFragment.getScene() != null)
+            prevFragment.getScene().transitioner = new SharedElementTransitioner(prevFragment, getSharedElementSet(stack.oldSharedElementNames));
         fragmentManager.popBackStack(String.valueOf(crumb), 0);
     }
 
@@ -33,7 +43,8 @@ class FragmentNavigator extends SceneNavigator {
             if (nextCrumb > 0) {
                 String prevKey = stack.keys.getString(nextCrumb - 1);
                 SceneFragment prevFramgent = (SceneFragment) fragmentManager.findFragmentByTag(prevKey);
-                sharedElements = getSharedElements(currentCrumb, crumb, prevFramgent, stack);
+                if (prevFramgent != null)
+                    sharedElements = getSharedElements(currentCrumb, crumb, prevFramgent, stack);
             }
             if (sharedElements != null) {
                 for(Pair sharedElement : sharedElements) {
@@ -64,5 +75,47 @@ class FragmentNavigator extends SceneNavigator {
         fragmentTransaction.replace(stack.getChildAt(0).getId(), new SceneFragment(scene, null), key);
         fragmentTransaction.addToBackStack(String.valueOf(crumb));
         fragmentTransaction.commit();
+    }
+
+    private Pair[] getOldSharedElements(int currentCrumb, int crumb, SharedElementContainer sharedElementContainer, final NavigationStackView stack) {
+        final HashMap<String, View> oldSharedElementsMap = getSharedElementMap(sharedElementContainer.getScene());
+        final Pair[] oldSharedElements = currentCrumb - crumb == 1 ? getSharedElements(oldSharedElementsMap, stack.oldSharedElementNames) : null;
+        if (oldSharedElements != null && oldSharedElements.length != 0) {
+            sharedElementContainer.setEnterCallback(new SharedElementCallback() {
+                @Override
+                public void onMapSharedElements(List<String> names, Map<String, View> elements) {
+                    for(int i = 0; i < stack.oldSharedElementNames.size(); i++) {
+                        String name = stack.oldSharedElementNames.getString(i);
+                        if (oldSharedElementsMap.containsKey(name)) {
+                            View oldSharedElement = oldSharedElementsMap.get(name);
+                            elements.put(names.get(i), oldSharedElement);
+                        }
+                    }
+                }
+            });
+            return oldSharedElements;
+        }
+        return null;
+    }
+
+    private Pair[] getSharedElements(int currentCrumb, int crumb, SharedElementContainer sharedElementContainer, final NavigationStackView stack) {
+        final HashMap<String, View> sharedElementsMap = getSharedElementMap(sharedElementContainer.getScene());
+        final Pair[] sharedElements = crumb - currentCrumb == 1 ? getSharedElements(sharedElementsMap, stack.sharedElementNames) : null;
+        if (sharedElements != null && sharedElements.length != 0) {
+            sharedElementContainer.setExitCallback(new SharedElementCallback() {
+                @Override
+                public void onMapSharedElements(List<String> names, Map<String, View> elements) {
+                    for(int i = 0; i < names.size(); i++) {
+                        String mappedName = names.get(i);
+                        if (stack.oldSharedElementNames != null && stack.oldSharedElementNames.size() > i)
+                            mappedName = stack.oldSharedElementNames.getString(i);
+                        if (sharedElementsMap.containsKey(mappedName))
+                            elements.put(names.get(i), sharedElementsMap.get(mappedName));
+                    }
+                }
+            });
+            return sharedElements;
+        }
+        return null;
     }
 }
