+++ NavigationReactNative/src/android/app/src/main/java/com/navigation/reactnative/SharedElementContainer.java
@@ -7,6 +7,8 @@ import androidx.core.app.SharedElementCallback;
 interface SharedElementContainer {
     SceneView getScene();
 
+    boolean canAddTarget();
+
     void setEnterTransition(Transition transition);
 
     void setExitCallback(SharedElementCallback sharedElementCallback);
